requirements.txt
aws-cdk-lib==2.89.0
constructs>=10.0.0,<11.0.0
pydantic>=2.5.1
pydantic-settings>=2.1.0
